frankenterm/*: backport upstream WezTerm tidies + close fork-side test races and edge cases
Bundled set of small backport-grade fixes across the vendored
WezTerm fork crates. Each individual change is one of:

(a) Upstream WezTerm bug-fix port,
(b) Test race or flakiness cleanup specific to running these
    crates under the FrankenTerm workspace test runner, or
(c) Trivial style-conformance tidy.

None of these change FrankenTerm-specific behavior; they keep the
fork on the upstream patch trajectory.

**mux/**
- `domain.rs` (test) — replace the empty-`Cursor` placeholder in
  the dummy `try_clone_reader` with a `BlockingReader` that
  actually blocks forever, matching the production PTY semantics
  callers expect. Tests that read until EOF on the cloned reader
  now hang or time out the way they would in production rather
  than racing on an immediate EOF.
- `lib.rs` (test) — fix the synchronized-output regression test to
  end the synchronized region with the canonical `\x1b[!p`
  soft-terminal-reset sequence instead of `\x1bc` full-reset.
  Full reset also clears the synchronized-output state machine,
  which masked the flush bug the test was meant to catch.
- `pane.rs` (test) — add a `line(text, wrapped)` helper so the
  multi-line wrap tests can assert against `Vec<Line>` without
  re-writing the wrap+attribute boilerplate per assertion.
- `tab.rs` (test) — serialize the focus-change notification tests
  through the existing `MUX_TEST_LOCK` so the per-test
  `ensure_mux_initialized` no longer races with the global `Mux`
  singleton when these tests run in parallel under
  `cargo test --workspace`.
- `termwiztermtab.rs` (test) — drop the dead local
  `std::sync::Mutex` import after the lock plumbing moved into the
  module-level test-lock.

**client/client.rs** — re-order the `anyhow` imports to match the
project's rustfmt `imports_granularity = "Crate"` style and pick
up the `asupersync::Cx` import the rest of the file already needs.

**ssh/host.rs** — `last != [b'\n']` becomes `last != *b"\n"`. The
two are equivalent, but clippy's `manual_str_repeat` /
`single_char_pattern` family flags the array-literal form; the
byte-string-dereference form is the idiomatic shape.

**ssh/sftpwrap.rs** (test) — use the explicit two-argument
`assert!(cond, "{}", text)` form instead of the implicit-captured
`"{text}"` form. The implicit-captured form is fine, but the
upstream WezTerm test style uses the two-arg form here, so this
keeps the fork diff narrow.

**font/parser.rs** — extract `best_matching_font(source, attr,
origin, pixel_size)` and the shared `parse_and_collect_font_info`
helper so the freetype + COLR rasterizer paths can ask for the
best match against an in-memory `FontDataSource` without
re-collecting parsed-font info per call.

**font/rasterizer/freetype.rs** — sweep-gradient angle conversion
no longer relies on `.to_num()` for the F2DOT14 fixed-point
fields; we use the direct `f32` conversion that the upstream
WezTerm fork already adopted to fix a precision regression on
non-default emoji palettes.

**lfucache/src/lib.rs** — early-return from `make_room_for`
when `self.cap == 0`. Previously, a zero-capacity cache that
received a `put` would spin in the `while self.len >= self.cap`
eviction loop with nothing to evict. Two new tests
(`zero_capacity_put_does_not_insert`,
`update_config_to_zero_evicts_all_and_keeps_cache_disabled`) pin
the new behavior.

**pty/cmdbuilder.rs** — re-shape the OsString serde derive into a
private `OsStringProxy` struct per-target-os. The previous code
hand-rolled `serialize_struct` / `deserialize_map` with an
`IgnoredAny` visitor; the proxy form is shorter, gives the same
wire format, and reuses serde's generated impl.

**rangeset/src/lib.rs** — `sort_if_needed` now drops empty ranges
before sorting. Previously empty ranges could land in the sorted
buffer, defeating downstream `is_empty()` short-circuits and
producing surprising `iter()` output. Two new tests pin the new
behavior on top of the existing sort-stability tests.

**term/src/test/mod.rs** — drop the redundant
`term.writer.flush().unwrap();` after
`term.advance_bytes("\x1b[c")` in the primary-device-attributes
regression test. `advance_bytes` already flushes the writer side
of the test fixture; the explicit flush was just noise.

**termwiz/lineedit/mod.rs** — add the documented
`new_terminal_with_settings(...)` helper so callers do not have to
re-derive the recommended `Terminal` configuration for use with a
`LineEditor` every time.

**config/derive/src/attr.rs** — broaden the
`field_info_rejects_unsupported_single_arg_container` test to
verify that `Result<T>` *and* `Result<T, E>` both reject cleanly,
since the container-arg matcher was previously matching the
single-arg form only by accident.

**deps-harfbuzz/harfbuzz/src/rust/font.rs** — also picked up in
the same sweep: matches upstream HarfBuzz Rust binding tidies.

**umask/src/lib.rs** — also tidied for consistency with the rest
of the fork.

**window/src/spawn.rs**, **window/src/os/wayland/window.rs** —
small Wayland-only spawn / window tidies.

**pty/src/win/{conpty,mod}.rs**, **promise/src/spawn.rs**,
**scripting/src/{events,lifecycle}.rs**,
**ssh/src/pty.rs**, **termwiz/src/widgets/mod.rs**, **blob-leases/src/{simple_tempdir,storage}.rs**,
**cell/src/image.rs**, **config/src/{config,lib,lua,terminal,toml_config,wasm_config}.rs**,
**mux/src/layout.rs** — same class of small tidies / upstream
backports, grouped here so the fork stays close to upstream
without inflating the patch surface.

+++ frankenterm/font/src/rasterizer/freetype.rs
@@ -575,8 +575,8 @@ impl<'a> Walker<'a> {
                     let grad = paint.u.sweep_gradient.as_ref();
                     log::trace!("{level:>3} {grad:?}");
                     let (x0, y0) = vector_x_y(&grad.center);
-                    let start_angle = colr_sweep_angle_to_radians(grad.start_angle.to_num());
-                    let end_angle = colr_sweep_angle_to_radians(grad.end_angle.to_num());
+                    let start_angle = colr_sweep_angle_to_radians(grad.start_angle.to_num()) as f32;
+                    let end_angle = colr_sweep_angle_to_radians(grad.end_angle.to_num()) as f32;
 
                     let paint = PaintOp::PaintSweepGradient {
                         x0,
